requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
dataclasses==0.7
Django==3.0.5
django-cors-headers==3.5.0
django-filter==2.3.0
django-heroku==0.3.1
djangorestframework==3.11.1
djongo==1.3.3
dnspython==1.16.0
gunicorn==20.0.4
heroku==0.1.4
idna==2.10
psycopg2==2.8.6
pymongo==3.11.0
python-dateutil==1.5
pytz==2020.1
requests==2.24.0
sqlparse==0.2.4
urllib3==1.25.10
